+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Forward portable activities from outbox to actors listed in `to` and `cc`.
 - Forward portable activities from outbox to other actor's clones.
 - Implement S2S inbox endpoints for portable users.
+- Added `pleroma.quote` property to Mastodon API `Status` entity.
 
 ### Fixed